+++ src/main/java/be/kdg/sa/velo/events/vehicles/locks/LockUndockedVehicleEvent.java
@@ -9,7 +9,7 @@ public class LockUndockedVehicleEvent extends LockVehicleEvent {
 	private final double latitude;
 	private final double longitude;
 	
-	protected LockUndockedVehicleEvent (long vehicleId, long userId, boolean defect, double latitude, double longitude) {
+	protected LockUndockedVehicleEvent (int vehicleId, int userId, boolean defect, double latitude, double longitude) {
 		super (vehicleId, userId, defect);
 		this.latitude = latitude;
 		this.longitude = longitude;
